requirements.txt
flask
flask-SQLAlchemy
python-dotenv
google-generativeai
chromadb
sentence-transformers
numpy
requests